requirements.txt
asgiref==3.8.1
crispy-bootstrap4==2024.1
Django==4.2.13
django-crispy-forms==2.1
django-filter==24.2
djangorestframework==3.15.1
importlib_metadata==7.1.0
Markdown==3.6
psycopg2==2.9.9
python-dotenv==1.0.1
sqlparse==0.5.0
typing_extensions==4.12.0
zipp==3.19.0
